requirements.txt
openai
selenium
beautifulsoup4
requests
python-docx
python-dotenv
webdriver-manager
flask